Initial addition of metrics classes.
I've inlcuded a simple interface that allows you to record an event, as well as start and stop a timer.
Stopping a timer submits the timer data.

There is a single implementation that allows for multiple backends.  Two backends
are included in this checkin.  One simply logs the data, the other forwards over UDP
to a statsd server.

These still need to be end to end tested.

I feel dirty excluding the impls from test coverage counts, but I'm not
sure how to test them.